Split new subsonic api converting extension functions into different files.
