+++ ultrasonic/src/main/kotlin/org/moire/ultrasonic/data/APIMusicFolderConverter.kt
@@ -1,0 +1,12 @@
+// Converts MusicFolder entity from [org.moire.ultrasonic.api.subsonic.SubsonicAPIClient]
+// to app domain entities.
+@file:JvmName("APIMusicFolderConverter")
+package org.moire.ultrasonic.data
+
+import org.moire.ultrasonic.domain.MusicFolder
+import org.moire.ultrasonic.api.subsonic.models.MusicFolder as APIMusicFolder
+
+fun APIMusicFolder.toDomainEntity(): MusicFolder = MusicFolder(this.id.toString(), this.name)
+
+fun List<APIMusicFolder>.toDomainEntityList(): List<MusicFolder>
+        = this.map { it.toDomainEntity() }
