--- ultrasonic/src/main/kotlin/org/moire/ultrasonic/data/SubsonicAPIConverter.kt
@@ -1,120 +1,0 @@
-// Converts entities from [org.moire.ultrasonic.api.subsonic.SubsonicAPIClient] to app domain entities.
-@file:JvmName("APIConverter")
-package org.moire.ultrasonic.data
-
-import org.moire.ultrasonic.api.subsonic.models.Album
-import org.moire.ultrasonic.api.subsonic.models.Index
-import org.moire.ultrasonic.api.subsonic.models.MusicDirectoryChild
-import org.moire.ultrasonic.api.subsonic.models.SearchThreeResult
-import org.moire.ultrasonic.api.subsonic.models.SearchTwoResult
-import org.moire.ultrasonic.domain.Artist
-import org.moire.ultrasonic.domain.Indexes
-import org.moire.ultrasonic.domain.MusicDirectory
-import org.moire.ultrasonic.domain.MusicFolder
-import org.moire.ultrasonic.domain.Playlist
-import org.moire.ultrasonic.domain.SearchResult
-import java.text.SimpleDateFormat
-import org.moire.ultrasonic.api.subsonic.models.Artist as APIArtist
-import org.moire.ultrasonic.api.subsonic.models.Indexes as APIIndexes
-import org.moire.ultrasonic.api.subsonic.models.MusicDirectory as APIMusicDirectory
-import org.moire.ultrasonic.api.subsonic.models.MusicFolder as APIMusicFolder
-import org.moire.ultrasonic.api.subsonic.models.Playlist as APIPlaylist
-import org.moire.ultrasonic.api.subsonic.models.SearchResult as APISearchResult
-
-fun APIMusicFolder.toDomainEntity(): MusicFolder = MusicFolder(this.id.toString(), this.name)
-
-fun List<APIMusicFolder>.toDomainEntityList(): List<MusicFolder>
-        = this.map { it.toDomainEntity() }
-
-fun APIIndexes.toDomainEntity(): Indexes = Indexes(this.lastModified, this.ignoredArticles,
-        this.shortcutList.map { it.toDomainEntity() }, this.indexList.foldIndexToArtistList())
-
-private fun List<Index>.foldIndexToArtistList(): List<Artist> = this.fold(listOf(), {
-    acc, index -> acc + index.artists.map { it.toDomainEntity() }
-})
-
-fun APIArtist.toDomainEntity(): Artist = Artist().apply {
-    id = this@toDomainEntity.id.toString()
-    name = this@toDomainEntity.name
-}
-
-fun APIArtist.toMusicDirectoryDomainEntity(): MusicDirectory = MusicDirectory().apply {
-    name = this@toMusicDirectoryDomainEntity.name
-    addAll(this@toMusicDirectoryDomainEntity.albumsList.map { it.toDomainEntity() })
-}
-
-fun Album.toDomainEntity(): MusicDirectory.Entry = MusicDirectory.Entry().apply {
-    id = this@toDomainEntity.id.toString()
-    setIsDirectory(true)
-    title = this@toDomainEntity.name
-    coverArt = this@toDomainEntity.coverArt
-    artist = this@toDomainEntity.artist
-    artistId = this@toDomainEntity.artistId.toString()
-    songCount = this@toDomainEntity.songCount.toLong()
-    duration = this@toDomainEntity.duration
-    created = this@toDomainEntity.created?.time
-    year = this@toDomainEntity.year
-    genre = this@toDomainEntity.genre
-}
-
-fun Album.toMusicDirectoryDomainEntity(): MusicDirectory = MusicDirectory().apply {
-    addAll(this@toMusicDirectoryDomainEntity.songList.map { it.toDomainEntity() })
-}
-
-fun MusicDirectoryChild.toDomainEntity(): MusicDirectory.Entry = MusicDirectory.Entry().apply {
-    id = this@toDomainEntity.id.toString()
-    parent = this@toDomainEntity.parent.toString()
-    setIsDirectory(this@toDomainEntity.isDir)
-    title = this@toDomainEntity.title
-    album = this@toDomainEntity.album
-    albumId = this@toDomainEntity.albumId.toString()
-    artist = this@toDomainEntity.artist
-    artistId = this@toDomainEntity.artistId.toString()
-    track = this@toDomainEntity.track
-    year = this@toDomainEntity.year
-    genre = this@toDomainEntity.genre
-    contentType = this@toDomainEntity.contentType
-    suffix = this@toDomainEntity.suffix
-    transcodedContentType = this@toDomainEntity.transcodedContentType
-    transcodedSuffix = this@toDomainEntity.transcodedSuffix
-    coverArt = this@toDomainEntity.coverArt
-    size = this@toDomainEntity.size
-    duration = this@toDomainEntity.duration
-    bitRate = this@toDomainEntity.bitRate
-    path = this@toDomainEntity.path
-    setIsVideo(this@toDomainEntity.isVideo)
-    created = this@toDomainEntity.created?.time
-    starred = this@toDomainEntity.starred != null
-    discNumber = this@toDomainEntity.discNumber
-    type = this@toDomainEntity.type
-}
-
-fun APIMusicDirectory.toDomainEntity(): MusicDirectory = MusicDirectory().apply {
-    name = this@toDomainEntity.name
-    addAll(this@toDomainEntity.childList.map { it.toDomainEntity() })
-}
-
-fun APISearchResult.toDomainEntity(): SearchResult = SearchResult(emptyList(), emptyList(),
-        this.matchList.map { it.toDomainEntity() })
-
-fun SearchTwoResult.toDomainEntity(): SearchResult = SearchResult(
-        this.artistList.map { it.toDomainEntity() },
-        this.albumList.map { it.toDomainEntity() },
-        this.songList.map { it.toDomainEntity() })
-
-fun SearchThreeResult.toDomainEntity(): SearchResult = SearchResult(
-        this.artistList.map { it.toDomainEntity() },
-        this.albumList.map { it.toDomainEntity() },
-        this.songList.map { it.toDomainEntity() })
-
-fun APIPlaylist.toMusicDirectoryDomainEntity(): MusicDirectory = MusicDirectory().apply {
-    name = this@toMusicDirectoryDomainEntity.name
-    addAll(this@toMusicDirectoryDomainEntity.entriesList.map { it.toDomainEntity() })
-}
-
-fun APIPlaylist.toDomainEntity(): Playlist = Playlist(this.id.toString(), this.name, this.owner,
-        this.comment, this.songCount.toString(),
-        this.created?.let { SimpleDateFormat.getDateTimeInstance().format(it.time) },
-        public.toString())
-
-fun List<APIPlaylist>.toDomainEntitiesList(): List<Playlist> = this.map { it.toDomainEntity() }
